Convert to typescript
Related to #5

Convert the codebase to TypeScript and add interfaces for each class.

* Rename `gameEngine.mjs` to `gameEngine.ts` and convert to TypeScript.
  - Add interfaces for `Player`, `Monster`, and `GameState`.
  - Update methods to use TypeScript types.
* Rename `index.mjs` to `index.ts` and convert to TypeScript.
  - Update import statements to use `.ts` extensions.
  - Add type annotations for `engine`, `renderer`, and `state` variables.
* Rename `renderer.mjs` to `renderer.ts` and convert to TypeScript.
  - Add interfaces for `Renderer` and `GameState` classes.
  - Update methods to use TypeScript types.
* Update `.vscode/launch.json` to reference `index.ts` instead of `index.mjs`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/pagannon/RG/issues/5?shareId=XXXX-XXXX-XXXX-XXXX).
