+++ .vscode/launch.json
@@ -11,7 +11,7 @@
             "skipFiles": [
                 "<node_internals>/**"
             ],
-            "program": "${workspaceFolder}\\index.mjs",
+            "program": "${workspaceFolder}\\index.ts",
             "console": "externalTerminal"
         }
     ]
